requirements.txt
zendriver==0.13.1
markdownify==1.2.0
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
openai==1.43.0
pydantic==2.9.2
Crawl4AI==0.7.3
fastmcp==2.11.3
fastapi==0.115.0
uvicorn==0.30.6